# productCenter/api/productmedia/v1/product_media_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v7.35.0
// source: productmedia/v1/product_media.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationProductMediaCreateMedia = "/productmedia.v1.ProductMedia/CreateMedia"
const OperationProductMediaDeleteMedia = "/productmedia.v1.ProductMedia/DeleteMedia"
const OperationProductMediaGetMedia = "/productmedia.v1.ProductMedia/GetMedia"
const OperationProductMediaListMedias = "/productmedia.v1.ProductMedia/ListMedias"
const OperationProductMediaUpdateMedia = "/productmedia.v1.ProductMedia/UpdateMedia"

type ProductMediaHTTPServer interface {
	// CreateMedia 上传副图（创建）
	CreateMedia(context.Context, *CreateMediaRequest) (*CreateMediaResponse, error)
	// DeleteMedia 删除副图
	DeleteMedia(context.Context, *DeleteMediaRequest) (*DeleteMediaResponse, error)
	// GetMedia 查看副图详情
	GetMedia(context.Context, *GetMediaRequest) (*GetMediaResponse, error)
	// ListMedias 分页查询副图列表（支持按商品筛选）
	ListMedias(context.Context, *ListMediasRequest) (*ListMediasResponse, error)
	// UpdateMedia 更新副图
	UpdateMedia(context.Context, *UpdateMediaRequest) (*UpdateMediaResponse, error)
}

func RegisterProductMediaHTTPServer(s *http.Server, srv ProductMediaHTTPServer) {
	r := s.Route("/")
	r.POST("/api/v1/product-media", _ProductMedia_CreateMedia0_HTTP_Handler(srv))
	r.PUT("/api/v1/product-media/{id}", _ProductMedia_UpdateMedia0_HTTP_Handler(srv))
	r.GET("/api/v1/product-media/{id}", _ProductMedia_GetMedia0_HTTP_Handler(srv))
	r.GET("/api/v1/product-media", _ProductMedia_ListMedias0_HTTP_Handler(srv))
	r.DELETE("/api/v1/product-media/{id}", _ProductMedia_DeleteMedia0_HTTP_Handler(srv))
}

func _ProductMedia_CreateMedia0_HTTP_Handler(srv ProductMediaHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateMediaRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationProductMediaCreateMedia)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateMedia(ctx, req.(*CreateMediaRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateMediaResponse)
		return ctx.Result(200, reply)
	}
}

func _ProductMedia_UpdateMedia0_HTTP_Handler(srv ProductMediaHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateMediaRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationProductMediaUpdateMedia)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateMedia(ctx, req.(*UpdateMediaRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateMediaResponse)
		return ctx.Result(200, reply)
	}
}

func _ProductMedia_GetMedia0_HTTP_Handler(srv ProductMediaHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetMediaRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationProductMediaGetMedia)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetMedia(ctx, req.(*GetMediaRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetMediaResponse)
		return ctx.Result(200, reply)
	}
}

func _ProductMedia_ListMedias0_HTTP_Handler(srv ProductMediaHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListMediasRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationProductMediaListMedias)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListMedias(ctx, req.(*ListMediasRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListMediasResponse)
		return ctx.Result(200, reply)
	}
}

func _ProductMedia_DeleteMedia0_HTTP_Handler(srv ProductMediaHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteMediaRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationProductMediaDeleteMedia)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteMedia(ctx, req.(*DeleteMediaRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteMediaResponse)
		return ctx.Result(200, reply)
	}
}

type ProductMediaHTTPClient interface {
	// CreateMedia 上传副图（创建）
	CreateMedia(ctx context.Context, req *CreateMediaRequest, opts ...http.CallOption) (rsp *CreateMediaResponse, err error)
	// DeleteMedia 删除副图
	DeleteMedia(ctx context.Context, req *DeleteMediaRequest, opts ...http.CallOption) (rsp *DeleteMediaResponse, err error)
	// GetMedia 查看副图详情
	GetMedia(ctx context.Context, req *GetMediaRequest, opts ...http.CallOption) (rsp *GetMediaResponse, err error)
	// ListMedias 分页查询副图列表（支持按商品筛选）
	ListMedias(ctx context.Context, req *ListMediasRequest, opts ...http.CallOption) (rsp *ListMediasResponse, err error)
	// UpdateMedia 更新副图
	UpdateMedia(ctx context.Context, req *UpdateMediaRequest, opts ...http.CallOption) (rsp *UpdateMediaResponse, err error)
}

type ProductMediaHTTPClientImpl struct {
	cc *http.Client
}

func NewProductMediaHTTPClient(client *http.Client) ProductMediaHTTPClient {
	return &ProductMediaHTTPClientImpl{client}
}

// CreateMedia 上传副图（创建）
func (c *ProductMediaHTTPClientImpl) CreateMedia(ctx context.Context, in *CreateMediaRequest, opts ...http.CallOption) (*CreateMediaResponse, error) {
	var out CreateMediaResponse
	pattern := "/api/v1/product-media"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationProductMediaCreateMedia))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// DeleteMedia 删除副图
func (c *ProductMediaHTTPClientImpl) DeleteMedia(ctx context.Context, in *DeleteMediaRequest, opts ...http.CallOption) (*DeleteMediaResponse, error) {
	var out DeleteMediaResponse
	pattern := "/api/v1/product-media/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationProductMediaDeleteMedia))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetMedia 查看副图详情
func (c *ProductMediaHTTPClientImpl) GetMedia(ctx context.Context, in *GetMediaRequest, opts ...http.CallOption) (*GetMediaResponse, error) {
	var out GetMediaResponse
	pattern := "/api/v1/product-media/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationProductMediaGetMedia))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListMedias 分页查询副图列表（支持按商品筛选）
func (c *ProductMediaHTTPClientImpl) ListMedias(ctx context.Context, in *ListMediasRequest, opts ...http.CallOption) (*ListMediasResponse, error) {
	var out ListMediasResponse
	pattern := "/api/v1/product-media"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationProductMediaListMedias))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateMedia 更新副图
func (c *ProductMediaHTTPClientImpl) UpdateMedia(ctx context.Context, in *UpdateMediaRequest, opts ...http.CallOption) (*UpdateMediaResponse, error) {
	var out UpdateMediaResponse
	pattern := "/api/v1/product-media/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationProductMediaUpdateMedia))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}